requirements.txt
apns2==0.7.2
asgiref==3.2.7
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
botocore==1.15.38
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
Django==2.1.15
django-choices==1.7.1
django-colorfield==0.3.2
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-health-check==3.12.1
django-jazzmin==2.2.10
django-money==1.1
django-otp==0.9.1
django-push-notifications==2.0.0
django-pyodbc-azure==2.1.0.0
django-split-settings==1.0.1
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-access-policy==0.5.1
drf-nested-routers==0.91
drf-yasg==1.17.1
Faker==4.4.0
fcm-django==0.3.7
graphviz==0.14
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lxml==4.6.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.3
Pillow==7.1.2
py-moneyed==0.8.0
pycparser==2.20
pydotplus==2.0.2
pyfcm==1.4.9
PyJWT==1.7.1
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
pywebpush==1.11.0
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
